only admin and owner can delete comments

+++ laravue-bug-tracker/app/Http/Controllers/Api/CommentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Comment;
 use App\Http\Controllers\Controller;
+use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -28,9 +29,15 @@ class CommentController extends Controller
 
     public function destroy($id)
     {
+        $user = User::find(Auth::user()->id);
         $comment = Comment::findOrFail($id);
-        $comment->delete();
 
-        return response()->json("success", 200);
+        if ($user->id === $comment->user_id || $user->hasAnyRole('admin')) {
+            $comment->delete();
+            return response()->json("success", 200);
+
+        }
+        return response()->json("failed", 401);
+
     }
 }
